Sleep friendly XBee sending.
This also adds a much longer delay (6hrs) and during that delay it
repeatedly tries to send the message until it’s successful.

At startup we also configure sleep mode to sleep for a few minutes at a
time before waking up. This way it’s never a few minutes after a sample
is taken that it’s transmitted.