Working on the Feather M0!
I'm trying to keep things as cross platform as possible just in case
we would like to toy with Arduino more, or anybody does for that matter.

Some lessons learned:

1) Before we would setup() the Serial's right before we used them. This
wouldn't work with the Serial2 for Conductivity because the calls to
pinPeripheral have to happen after, even if we had done them before. They
seem to get reset? I'm guessing the calls I see in Arduino-SAMD are
using the wrong pads and that's what makes order important.

2) Types are fun because Arduino uses SoftwareSerial and we use Uart.

3) Sleep wasn't working when I moved, I assume because there's buffering
happening with the hardware Uart that wasn't with the SoftwareSerial
and so it would happen too fast. Thankfully we are supposed to get an *OK
back from the Atlas boards and so I just wait for that now.