Buffer packets in memory. Things were crashing...
So I'm worried that the crash I was seeing is happening to either the
sensor or the collector. I have never seen this though. Perhaps an IRQ
firing while we're trying to do SD stuff? I've increased the rate of atlas
transmissions locally to see if I can get things to break with that
involved. Until then, this seems to be working great.