Leave RockBlock on forever after power up, but leave off while charging.
Kept seeing a weird issue in the field where we'd be unable to talk to the
RB sometimes, though the first few times would be fine. I think this
has something to do with longer interals of transmission, like 6 hours or
so and the capacitor can discharge and leave the RB in a funky state.

What we do now to try and avoid that is only power on the RB after we've
done our initial pause for charge, which on healthy restarts should be fast
and otherwise just leave the RB connected. Our worry isn't the RB draining
us but just keeping us from "turning on" which this should prevent.