Save defaults right away, just in case we don't later.
This fixes a bug where if you start with no memory card right after
programming them, they will end up with "new" memory that doesn't
look uninitialized and has all 1's. The problem is then that we don't
clear that schedule later when a memory card is inserted and we just
trust it, as we should. This saves defaults right away.